+++ src/HLS/binding/interconnection/mux_connection_binding.hpp
@@ -129,12 +129,12 @@ class mux_connection_binding : public conn_binding_creator
     */
    void determine_connection(const vertex& op, const HLS_manager::io_binding_type& var, generic_objRef fu_obj,
                              unsigned int port_num, unsigned int port_index, const OpGraphConstRef data,
-                             unsigned int precision, unsigned int alignment = 0);
+                             unsigned long long precision, unsigned int alignment = 0);
 
    /**
     * Compute the bitsize given a io_binding type
     */
-   unsigned int object_bitsize(const tree_managerRef TreeM, const HLS_manager::io_binding_type& obj) const;
+   unsigned long long object_bitsize(const tree_managerRef TreeM, const HLS_manager::io_binding_type& obj) const;
 
    /**
     * Recursive function which returns the offset of a dynamic multidimensional array call
@@ -142,7 +142,7 @@ class mux_connection_binding : public conn_binding_creator
    generic_objRef dynamic_multidimensional_array_handler(array_ref* ar, const vertex& op, const OpGraphConstRef data,
                                                          unsigned int& base_address_index_pointer,
                                                          std::vector<unsigned int>& recursive_indexes_values,
-                                                         std::vector<unsigned int>& dims, generic_objRef& global_adder,
+                                                         std::vector<unsigned long long> &dims, generic_objRef& global_adder,
                                                          const bool is_not_a_phi);
 
    /**
@@ -150,7 +150,7 @@ class mux_connection_binding : public conn_binding_creator
     * type
     */
    void connect_array_index(unsigned int tree_index, generic_objRef fu_obj, unsigned int port_num,
-                            unsigned int port_index, unsigned int bus_addr_bitsize, const OpGraphConstRef data,
+                            unsigned int port_index, unsigned long long bus_addr_bitsize, const OpGraphConstRef data,
                             const vertex& op);
 
  private:
@@ -159,24 +159,24 @@ class mux_connection_binding : public conn_binding_creator
     */
    void create_single_conn(const OpGraphConstRef data, const vertex& op, generic_objRef fu_obj_src,
                            generic_objRef fu_obj, unsigned int port_num, unsigned int port_index, unsigned int tree_var,
-                           unsigned int precision, const bool is_not_a_phi);
+                           unsigned long long precision, const bool is_not_a_phi);
 
    /**
     * connect the fu_obj with the associated registers.
     */
    void connect_to_registers(vertex op, const OpGraphConstRef data, generic_objRef fu_obj, unsigned int port_num,
-                             unsigned int port_index, unsigned int tree_var, unsigned int precision,
+                             unsigned int port_index, unsigned int tree_var, unsigned long long precision,
                              const bool is_not_a_phi);
 
    unsigned int extract_parm_decl(unsigned int tree_var, const tree_managerRef TreeM);
 
-   void add_conversion(unsigned int num, unsigned int size_tree_var, VertexIterator op, unsigned int form_par_type,
+   void add_conversion(unsigned int num, unsigned long long size_tree_var, VertexIterator op, unsigned int form_par_type,
                        unsigned int port_index, const generic_objRef fu_obj, const OpGraphConstRef data,
                        const tree_managerRef TreeM, unsigned int tree_var,
-                       const std::vector<HLS_manager::io_binding_type>& var_read, unsigned int size_form_par);
+                       const std::vector<HLS_manager::io_binding_type>& var_read, unsigned long long size_form_par);
 
-   unsigned int address_precision(unsigned int precision, const vertex& op, const OpGraphConstRef data,
-                                  const tree_managerRef TreeM);
+   unsigned long long address_precision(unsigned long long precision, const vertex& op, const OpGraphConstRef data,
+                                        const tree_managerRef TreeM);
 
    bool isZeroObj(unsigned int tree_index, const tree_managerRef TreeM);
    bool isConstantObj(unsigned int tree_index, const tree_managerRef TreeM);
