+++ src/HLS/binding/register/reg_binding.hpp
@@ -87,7 +87,7 @@ class reg_binding : public variable2obj<generic_objRef>
    std::map<unsigned int, unsigned int> reverse_map;
 
    /// relation between registers and their bitsize
-   std::map<unsigned int, unsigned int> bitsize_map;
+   std::map<unsigned int, unsigned long long> bitsize_map;
 
    /// HLS datastructure
    hlsRef HLS;
@@ -205,7 +205,7 @@ class reg_binding : public variable2obj<generic_objRef>
    /**
     * return bitsize
     */
-   unsigned int get_bitsize(unsigned int r) const;
+   unsigned long long get_bitsize(unsigned int r) const;
 
  private:
    /**
@@ -220,7 +220,7 @@ class reg_binding : public variable2obj<generic_objRef>
     * @param r is the register
     * @return the bitsize of register r
     */
-   unsigned int compute_bitsize(unsigned int r);
+   unsigned long long compute_bitsize(unsigned int r);
 };
 
 /**
