+++ src/HLS/virtual_components/conv_conn_obj.hpp
@@ -52,7 +52,7 @@
 class uu_conv_conn_obj : public generic_obj
 {
    /// number of bit of in/out ports
-   unsigned int bitsize;
+   unsigned long long bitsize;
 
  public:
    /**
@@ -70,7 +70,7 @@ class uu_conv_conn_obj : public generic_obj
    /**
     * add a size to the component
     */
-   void add_bitsize(unsigned int _bitsize)
+   void add_bitsize(unsigned long long _bitsize)
    {
       bitsize = _bitsize > bitsize ? _bitsize : bitsize;
    }
@@ -78,7 +78,7 @@ class uu_conv_conn_obj : public generic_obj
    /**
     * return the maximum bitsize associated with the component
     */
-   unsigned int get_bitsize() const
+   unsigned long long get_bitsize() const
    {
       return bitsize;
    }
@@ -91,7 +91,7 @@ class uu_conv_conn_obj : public generic_obj
 class ui_conv_conn_obj : public generic_obj
 {
    /// number of bit of in/out ports
-   unsigned int bitsize;
+   unsigned long long bitsize;
 
  public:
    /**
@@ -109,7 +109,7 @@ class ui_conv_conn_obj : public generic_obj
    /**
     * add a size to the component
     */
-   void add_bitsize(unsigned int _bitsize)
+   void add_bitsize(unsigned long long _bitsize)
    {
       bitsize = _bitsize > bitsize ? _bitsize : bitsize;
    }
@@ -117,7 +117,7 @@ class ui_conv_conn_obj : public generic_obj
    /**
     * return the maximum bitsize associated with the component
     */
-   unsigned int get_bitsize() const
+   unsigned long long get_bitsize() const
    {
       return bitsize;
    }
@@ -130,7 +130,7 @@ class ui_conv_conn_obj : public generic_obj
 class iu_conv_conn_obj : public generic_obj
 {
    /// number of bit of in/out ports
-   unsigned int bitsize;
+   unsigned long long bitsize;
 
  public:
    /**
@@ -148,7 +148,7 @@ class iu_conv_conn_obj : public generic_obj
    /**
     * add a size to the component
     */
-   void add_bitsize(unsigned int _bitsize)
+   void add_bitsize(unsigned long long _bitsize)
    {
       bitsize = _bitsize > bitsize ? _bitsize : bitsize;
    }
@@ -156,7 +156,7 @@ class iu_conv_conn_obj : public generic_obj
    /**
     * return the maximum bitsize associated with the component
     */
-   unsigned int get_bitsize() const
+   unsigned long long get_bitsize() const
    {
       return bitsize;
    }
@@ -169,7 +169,7 @@ class iu_conv_conn_obj : public generic_obj
 class ii_conv_conn_obj : public generic_obj
 {
    /// number of bit of in/out ports
-   unsigned int bitsize;
+   unsigned long long bitsize;
 
  public:
    /**
@@ -187,7 +187,7 @@ class ii_conv_conn_obj : public generic_obj
    /**
     * add a size to the component
     */
-   void add_bitsize(unsigned int _bitsize)
+   void add_bitsize(unsigned long long _bitsize)
    {
       bitsize = _bitsize > bitsize ? _bitsize : bitsize;
    }
@@ -195,7 +195,7 @@ class ii_conv_conn_obj : public generic_obj
    /**
     * return the maximum bitsize associated with the component
     */
-   unsigned int get_bitsize() const
+   unsigned long long get_bitsize() const
    {
       return bitsize;
    }
@@ -208,9 +208,9 @@ class ii_conv_conn_obj : public generic_obj
 class ff_conv_conn_obj : public generic_obj
 {
    /// number of bit of in ports
-   unsigned int bitsize_in;
+   unsigned long long bitsize_in;
    /// number of bit of out ports
-   unsigned int bitsize_out;
+   unsigned long long bitsize_out;
 
  public:
    /**
@@ -228,7 +228,7 @@ class ff_conv_conn_obj : public generic_obj
    /**
     * add a size in to the component
     */
-   void add_bitsize_in(unsigned int _bitsize)
+   void add_bitsize_in(unsigned long long _bitsize)
    {
       bitsize_in = _bitsize > bitsize_in ? _bitsize : bitsize_in;
    }
@@ -236,7 +236,7 @@ class ff_conv_conn_obj : public generic_obj
    /**
     * add a size in to the component
     */
-   void add_bitsize_out(unsigned int _bitsize)
+   void add_bitsize_out(unsigned long long _bitsize)
    {
       bitsize_out = _bitsize > bitsize_out ? _bitsize : bitsize_out;
    }
@@ -244,14 +244,14 @@ class ff_conv_conn_obj : public generic_obj
    /**
     * return the input bitsize associated with the component
     */
-   unsigned int get_bitsize_in() const
+   unsigned long long get_bitsize_in() const
    {
       return bitsize_in;
    }
    /**
     * return the output bitsize associated with the component
     */
-   unsigned int get_bitsize_out() const
+   unsigned long long get_bitsize_out() const
    {
       return bitsize_out;
    }
@@ -264,7 +264,7 @@ class ff_conv_conn_obj : public generic_obj
 class i_assign_conn_obj : public generic_obj
 {
    /// number of bit of in/out ports
-   unsigned int bitsize;
+   unsigned long long bitsize;
 
  public:
    /**
@@ -282,7 +282,7 @@ class i_assign_conn_obj : public generic_obj
    /**
     * add a size to the component
     */
-   void add_bitsize(unsigned int _bitsize)
+   void add_bitsize(unsigned long long _bitsize)
    {
       bitsize = _bitsize > bitsize ? _bitsize : bitsize;
    }
@@ -290,7 +290,7 @@ class i_assign_conn_obj : public generic_obj
    /**
     * return the maximum bitsize associated with the component
     */
-   unsigned int get_bitsize() const
+   unsigned long long get_bitsize() const
    {
       return bitsize;
    }
@@ -303,7 +303,7 @@ class i_assign_conn_obj : public generic_obj
 class u_assign_conn_obj : public generic_obj
 {
    /// number of bit of in/out ports
-   unsigned int bitsize;
+   unsigned long long bitsize;
 
  public:
    /**
@@ -321,7 +321,7 @@ class u_assign_conn_obj : public generic_obj
    /**
     * add a size to the component
     */
-   void add_bitsize(unsigned int _bitsize)
+   void add_bitsize(unsigned long long _bitsize)
    {
       bitsize = _bitsize > bitsize ? _bitsize : bitsize;
    }
@@ -329,7 +329,7 @@ class u_assign_conn_obj : public generic_obj
    /**
     * return the maximum bitsize associated with the component
     */
-   unsigned int get_bitsize() const
+   unsigned long long get_bitsize() const
    {
       return bitsize;
    }
@@ -342,7 +342,7 @@ class u_assign_conn_obj : public generic_obj
 class vb_assign_conn_obj : public generic_obj
 {
    /// number of bit of in/out ports
-   unsigned int bitsize;
+   unsigned long long bitsize;
 
  public:
    /**
@@ -360,7 +360,7 @@ class vb_assign_conn_obj : public generic_obj
    /**
     * add a size to the component
     */
-   void add_bitsize(unsigned int _bitsize)
+   void add_bitsize(unsigned long long _bitsize)
    {
       bitsize = _bitsize > bitsize ? _bitsize : bitsize;
    }
@@ -368,7 +368,7 @@ class vb_assign_conn_obj : public generic_obj
    /**
     * return the maximum bitsize associated with the component
     */
-   unsigned int get_bitsize() const
+   unsigned long long get_bitsize() const
    {
       return bitsize;
    }
@@ -381,7 +381,7 @@ class vb_assign_conn_obj : public generic_obj
 class f_assign_conn_obj : public generic_obj
 {
    /// number of bit of in/out ports
-   unsigned int bitsize;
+   unsigned long long bitsize;
 
  public:
    /**
@@ -399,7 +399,7 @@ class f_assign_conn_obj : public generic_obj
    /**
     * add a size to the component
     */
-   void add_bitsize(unsigned int _bitsize)
+   void add_bitsize(unsigned long long _bitsize)
    {
       bitsize = _bitsize > bitsize ? _bitsize : bitsize;
    }
@@ -407,7 +407,7 @@ class f_assign_conn_obj : public generic_obj
    /**
     * return the maximum bitsize associated with the component
     */
-   unsigned int get_bitsize() const
+   unsigned long long get_bitsize() const
    {
       return bitsize;
    }
